move the WireAppState out ouf Main.kt
