+++ app/src/main/kotlin/com/wire/android/ui/main/Main.kt
@@ -2,29 +2,22 @@ package com.wire.android.ui.main
 
 import androidx.compose.animation.ExperimentalAnimationApi
 import androidx.compose.material.Scaffold
-import androidx.compose.material.ScaffoldState
-import androidx.compose.material.rememberScaffoldState
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import androidx.navigation.compose.rememberNavController
 import com.wire.android.ui.main.archive.ArchiveScreen
 import com.wire.android.ui.main.conversation.ConversationRoute
 import com.wire.android.ui.main.conversation.ConversationViewModel
 import com.wire.android.ui.main.navigation.MainDrawer
 import com.wire.android.ui.main.navigation.MainNavigationScreenItem
 import com.wire.android.ui.main.navigation.MainTopBar
-import com.wire.android.ui.main.navigation.getCurrentNavigationItem
 import com.wire.android.ui.main.settings.SettingsScreen
 import com.wire.android.ui.main.support.SupportScreen
 import com.wire.android.ui.main.userprofile.UserProfileScreen
 import com.wire.android.ui.main.vault.VaultScreen
-import kotlinx.coroutines.CoroutineScope
+import com.wire.android.ui.rememberWireAppState
 
 @ExperimentalAnimationApi
 @ExperimentalComposeUiApi
@@ -37,7 +30,6 @@ fun MainScreen() {
         topBar = {
             MainTopBar(
                 wireAppState,
-                onTextChanged
             )
         },
         drawerContent = {
@@ -62,36 +54,10 @@ fun MainScreen() {
     }
 }
 
-
 @ExperimentalAnimationApi
 @ExperimentalComposeUiApi
 @Preview(showBackground = true)
 @Composable
 fun MainScreenPreview() {
     MainScreen()
-}
-
-@Composable
-fun rememberWireAppState(
-    scaffoldState: ScaffoldState = rememberScaffoldState(),
-    navController: NavHostController = rememberNavController(),
-    coroutineScope: CoroutineScope = rememberCoroutineScope()
-) = remember(scaffoldState, navController, coroutineScope) {
-    WireAppState(scaffoldState, navController, coroutineScope)
-}
-
-class WireAppState(
-    val scaffoldState: ScaffoldState,
-    val navController: NavHostController,
-    val coroutineScope: CoroutineScope,
-) {
-
-    val drawerState
-        get() = scaffoldState.drawerState
-
-    val currentNavigationItem
-        @Composable get() = navController.getCurrentNavigationItem()
-
-    val shouldShowSearchBar: Boolean
-        @Composable get() = currentNavigationItem?.hasSearchableTopBar ?: false
 }
