+++ app/src/main/kotlin/com/wire/android/ui/main/navigation/NavigationUtils.kt
@@ -8,30 +8,9 @@ import androidx.navigation.compose.currentBackStackEntryAsState
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 
-internal fun navigateToItem(
-    navController: NavController,
-    item: MainNavigationScreenItem,
-    scope: CoroutineScope,
-    scaffoldState: ScaffoldState
-) {
-    navController.navigate(item.route) {
-        navController.graph.startDestinationRoute?.let { route ->
-            popUpTo(route) {
-                saveState = true
-            }
-        }
-        launchSingleTop = true
-        restoreState = true
-    }
-    scope.launch { scaffoldState.drawerState.close() }
-}
-
 @Composable
 internal fun NavController.getCurrentNavigationItem(): MainNavigationScreenItem? {
     val navBackStackEntry by currentBackStackEntryAsState()
     val currentRoute = navBackStackEntry?.destination?.route
     return MainNavigationScreenItem.fromRoute(currentRoute)
 }
-
-@Composable
-internal fun NavController.isCurrentNavigationItemSearchable(): Boolean = getCurrentNavigationItem()?.hasSearchableTopBar ?: false
