+++ app/src/main/kotlin/com/wire/android/ui/main/vault/VaultScreen.kt
@@ -2,16 +2,16 @@ package com.wire.android.ui.main.vault
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
+import com.wire.android.ui.common.SearchBar
 import com.wire.android.ui.common.UnderConstructionScreen
 
 @Composable
-fun VaultScreen(onSearchTextChanged : (String) -> Unit) {
-
+fun VaultScreen() {
     UnderConstructionScreen(screenName = "VaultScreen")
 }
 
 @Preview(showBackground = false)
 @Composable
 fun VaultScreenPreview() {
-    //VaultScreen()
+    VaultScreen()
 }
